Refactor methods out of ApplicationController
Remove methods from ApplicationController and add them to new Encoder
and NotificationEndpoint concerns.
