+++ app/controllers/videos_controller.rb
@@ -3,6 +3,7 @@
 #
 # @see Video
 class VideosController < ApplicationController
+  include Encoder
   before_action :authenticate_admin!
 
   def index
@@ -34,7 +35,7 @@ class VideosController < ApplicationController
   end
 
   ##
-  # @see ApplicationController#create_media
+  # @see Encoder#create_media
   def create
     @video = Video.new(video_params)
     create_media(@video, 'video')
@@ -48,7 +49,7 @@ class VideosController < ApplicationController
   end
 
   ##
-  # @see ApplicationController#transcoding_notifications
+  # @see Encoder#transcoding_notifications
   def notifications_url
     Settings.app_scheme + Settings.zencoder.notification_user + ':' \
       + Settings.zencoder.notification_pass + '@' \
@@ -56,7 +57,7 @@ class VideosController < ApplicationController
   end
 
   ##
-  # @see ApplicationController#media_outputs
+  # @see Encoder#media_outputs
   def media_outputs
     Settings.video_outputs.map { |o| o.to_h }
   end
