+++ spec/controllers/audio_notifications_controller_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+describe AudioNotificationsController, type: :controller do
+  it_behaves_like 'a notification endpoint', Audio
+end
