+++ spec/controllers/audios_controller_spec.rb
@@ -7,6 +7,7 @@ describe AudiosController, type: :controller do
   let(:invalid_attributes) { attributes_for(:invalid_audio_factory) }
 
   it_behaves_like 'admin-only route', :index, :show, :new, :create
+  it_behaves_like 'an encoder', 'audio'
 
   context 'admin logged in' do
     login_admin
